Move harmony patch to separate project
