+++ Source/RealFuels.sln
@@ -4,6 +4,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 VisualStudioVersion = 16.0.31229.75
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RealFuels", "RealFuels.csproj", "{0041813D-DCD1-4AC7-8327-85765BF924A3}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBD}") = "RealFuelsB9PSPatch", "B9PSPatch/RealFuelsB9PSPatch.csproj", "{0041813D-DCD1-4AC7-8327-85765BF924A3}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
